+++ pages/client-dashboard.tsx
@@ -1,12 +1,12 @@
 import type { NextPage } from 'next'
-import { useParamClientId } from 'library/navigation';
+import { useParamId } from 'library/navigation';
 import ClientShow from 'components/views/client/ClientShow';
 import AuthPageWithStore from 'components/utility/AuthPageWithStore';
 import InvoiceTable from 'components/views/invoice/InvoiceTable';
 import Card from 'components/ui/layout/Card';
 
 const ClientDashboardPage: NextPage = () => {
-    const clientId = useParamClientId();
+    const clientId = useParamId();
 
     return <Card size='big' fullscreen={false} background="">
         <ClientShow clientId={clientId} />