+++ pages/invoice.tsx
@@ -1,19 +1,20 @@
 import type { NextPage } from 'next'
 import InvoiceEdition from 'components/views/invoice/InvoiceEdition';
-import { useGoBack, useGoInvoiceView, useParamInvoiceId } from 'library/navigation';
+import { useGoBack, useGoInvoiceView, useParamClientId, useParamId } from 'library/navigation';
 import AuthPageWithStore from 'components/utility/AuthPageWithStore';
 import { useCallback } from 'react';
 import { ClientInvoice } from 'models/Invoice';
 
 const InvoicePage: NextPage = () => {
-    const invoiceId = useParamInvoiceId();
+    const invoiceId = useParamId();
+    const clientId = useParamClientId();
     const goView = useGoInvoiceView();
     const goBack = useGoBack();
     const goViewSaved = useCallback((clientInvoice:ClientInvoice) => {
         goView(clientInvoice.invoice.id)
     }, [goView])
 
-    return <InvoiceEdition onCancel={goBack} onSave={goViewSaved} invoiceId={invoiceId} />
+    return <InvoiceEdition onCancel={goBack} onSave={goViewSaved} invoiceId={invoiceId} clientId={clientId} />
 }
 
 export default AuthPageWithStore(InvoicePage)