+++ changelog.md
@@ -22,6 +22,7 @@
 # 2022 03 23
 1. add latest clients in dashboard
 2. implement sorting and pagination of clients
+3. pre-fill client selector when creating an invoice from client profile
 
 # 2022 03 22
 - invalidate/modify pages when invoice is created.